cl-handle-unsupported-pubkey.patch
See 733a0a47b4b161785ab061aa17f230116466425a
If we receive an auth attempt with an unsupported public key type
(such as ec22519), unconditionally fallback to password auth instead
of the previous behaviour of diconnecting.

Much of the "password then pubkey" mentality is showing in these
bits of Cryptlib.